주석 작성, 캘린더 컴포넌트화
