+++ components/commons/filter/Filter.module.scss
@@ -1,7 +1,4 @@
 .filter {
-  position: absolute;
-  bottom: -10px;
-  left: 0;
   padding: 24px 20px;
   width: 390px;
   box-shadow: 0px 2px 8px 0px rgba(120, 116, 134, 0.25);
